requirements.txt
# requirements.txt
# Core Django
Django==5.1.7  # Updated for Python 3.14 compatibility
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.11

# Image Processing - Use version that supports Python 3.14
Pillow==11.3.0

# API Documentation
drf-yasg==1.21.8

# Environment Variables
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Production
gunicorn==22.0.0
whitenoise==6.9.0

# Testing
pytest==8.3.5
pytest-django==4.10.0